// frontend/src/api/dto/drink.dto.ts
class DrinkDto {
  id: string;
  category: string;
  name: string;
  location: string;
  description?: string;
  genre?: string;
  link: string;
  extraLink?: string;
  dateTime: string;
  latitude: string;
  longitude: string;
  picture?: string;
  priceRangeMin?: number;
  priceRangeMax?: number;

  constructor(data: Event) {
    this.id = data.fsq_id;
    this.category = 'eat';
    this.name = data?.name;
    this.location = data?.location?.locality;
    this.description = data?.description;
    this.genre = data?.categories?.[0]?.name; // TODO: convert to array of strings
    this.link = data?.link;
    this.extraLink = data.menu;
    this.dateTime = data.hours?.display;
    this.latitude = data?.geocodes?.main?.latitude.toString();
    this.longitude = data?.geocodes?.main?.longitude.toString();
    this.picture = data?.photos?.[0]?.prefix + data?.photos?.[0]?.suffix;
    this.priceRangeMin = data?.price;
    this.priceRangeMax = data?.price;
  }
}

export default DrinkDto;

export interface Event {
  fsq_id: string;
  categories: Category[];
  chains: Chain[];
  closed_bucket: string;
  date_closed: string;
  description: string;
  distance: number;
  email: string;
  fax: string;
  features: Features;
  geocodes: Geocodes;
  hours: Hours;
  hours_popular: HoursPopular[];
  link: string;
  location: Location;
  menu: string;
  name: string;
  photos: Photo[];
  popularity: number;
  price: number;
  rating: number;
  related_places: RelatedPlaces;
  social_media: SocialMedia;
  stats: Stats;
  store_id: string;
  tastes: string[];
  tel: string;
  timezone: string;
  tips: Tip3[];
  venue_reality_bucket: string;
  verified: boolean;
  website: string;
}

export interface Category {
  id: number;
  name: string;
  short_name: string;
  plural_name: string;
  icon: Icon;
}

export interface Icon {
  id: string;
  created_at: string;
  prefix: string;
  suffix: string;
  width: number;
  height: number;
  classifications: string[];
  tip: Tip;
}

export interface Tip {
  id: string;
  created_at: string;
  text: string;
  url: string;
  lang: string;
  agree_count: number;
  disagree_count: number;
}

export interface Chain {
  id: string;
  name: string;
}

export interface Features {
  payment: Payment;
  food_and_drink: FoodAndDrink;
  services: Services;
  amenities: Amenities;
  attributes: Attributes;
}
// ...
export interface Geocodes {
  drop_off: DropOff;
  front_door: FrontDoor;
  main: Main;
  road: Road;
  roof: Roof;
}

export interface DropOff {
  latitude: number;
  longitude: number;
}

export interface FrontDoor {
  latitude: number;
  longitude: number;
}

export interface Main {
  latitude: number;
  longitude: number;
}

export interface Road {
  latitude: number;
  longitude: number;
}

export interface Roof {
  latitude: number;
  longitude: number;
}

export interface Hours {
  display: string;
  is_local_holiday: boolean;
  open_now: boolean;
  regular: Regular[];
}

export interface Regular {
  close: string;
  day: number;
  open: string;
}

export interface HoursPopular {
  close: string;
  day: number;
  open: string;
}

export interface Location {
  address: string;
  address_extended: string;
  admin_region: string;
  census_block: string;
  country: string;
  cross_street: string;
  dma: string;
  formatted_address: string;
  locality: string;
  neighborhood: string[];
  po_box: string;
  post_town: string;
  postcode: string;
  region: string;
}

export interface Photo {
  id: string;
  created_at: string;
  prefix: string;
  suffix: string;
  width: number;
  height: number;
  classifications: string[];
  tip: Tip2;
}

export interface Tip2 {
  id: string;
  created_at: string;
  text: string;
  url: string;
  lang: string;
  agree_count: number;
  disagree_count: number;
}

export interface RelatedPlaces {}

export interface SocialMedia {
  facebook_id: string;
  instagram: string;
  twitter: string;
}

export interface Stats {
  total_photos: number;
  total_ratings: number;
  total_tips: number;
}

export interface Tip3 {
  id: string;
  created_at: string;
  text: string;
  url: string;
  lang: string;
  agree_count: number;
  disagree_count: number;
}